Updated Rude Edit Dialog
